requirements.txt
# Tkinter
Pillow
rpi_ws281x
adafruit-circuitpython-neopixel
